+++ integrations/github/github/core/options.py
@@ -39,3 +39,11 @@ class ListIssueOptions(RepositoryIdentifier):
     """Options for listing issues."""
 
     state: Required[str]
+
+
+class SingleUserOptions(TypedDict):
+    login: Required[str]
+
+
+class SingleTeamOptions(TypedDict):
+    slug: Required[str]