+++ integrations/github/github/webhook/events.py
@@ -10,6 +10,8 @@ REPOSITORY_UPSERT_EVENTS = [
 ]
 REPOSITORY_DELETE_EVENTS = ["archived", "deleted"]
 
+USER_UPSERT_EVENTS = ["member_added"]
+USER_DELETE_EVENTS = ["member_removed"]
 
 # Pull request events
 PULL_REQUEST_EVENTS = [
@@ -23,6 +25,12 @@ PULL_REQUEST_EVENTS = [
     "closed",
 ]
 
+TEAM_UPSERT_EVENTS = ["created", "edited"]
+TEAM_DELETE_EVENTS = ["deleted"]
+
+
+TEAM_EVENTS = TEAM_UPSERT_EVENTS + TEAM_DELETE_EVENTS
+USER_EVENTS = USER_UPSERT_EVENTS + USER_DELETE_EVENTS
 
 # Issue events
 ISSUE_UPSERT_EVENTS = [
@@ -53,7 +61,9 @@ ALL_EVENTS = (
     + REPOSITORY_DELETE_EVENTS
     + PULL_REQUEST_EVENTS
     + ISSUE_EVENTS
+    + TEAM_EVENTS
+    + USER_EVENTS
 )
 
 
-WEBHOOK_CREATE_EVENTS = ["repository", "pull_request", "issues"]
+WEBHOOK_CREATE_EVENTS = ["repository", "pull_request", "issues", "organization", "team"]