+++ integrations/github/CHANGELOG.md
@@ -38,6 +38,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped ocean version to ^0.24.7
 
+### Features
+
+- Implement support for User kinds
+- Implement support for Team kinds
+
 
 ## 0.1.10-dev (2025-06-09)
 
@@ -95,7 +100,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## 0.1.3-dev (2025-06-01)
-
 
 ### Improvements